+++ scripts/docs/promptflow.client.rst
@@ -1,0 +1,14 @@
+promptflow.client package
+==============================
+
+.. autoclass:: promptflow.client.PFClient
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :noindex:
+
+.. autoclass:: promptflow.client.load_run
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :noindex:
