+++ scripts/docs/promptflow.core.rst
@@ -1,0 +1,20 @@
+promptflow.core package
+==============================
+
+.. autoclass:: promptflow.core.log_metric
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :noindex:
+
+.. autoclass:: promptflow.core.ToolProvider
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :noindex:
+
+.. autoclass:: promptflow.core.tool
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :noindex:
