+++ src/promptflow/promptflow/_sdk/_constants.py
@@ -7,6 +7,13 @@ from collections import namedtuple
 from enum import Enum
 from pathlib import Path
 
+from promptflow._constants import (
+    CONNECTION_SCRUBBED_VALUE,
+    ConnectionAuthMode,
+    ConnectionType,
+    CustomStrongTypeConnectionConfigs,
+)
+
 LOGGER_NAME = "promptflow"
 
 PROMPT_FLOW_HOME_DIR_ENV_VAR = "PF_HOME_DIRECTORY"
@@ -104,7 +111,7 @@ KEYRING_ENCRYPTION_KEY_NAME = "encryption_key"
 KEYRING_ENCRYPTION_LOCK_PATH = (HOME_PROMPT_FLOW_DIR / "encryption_key.lock").resolve()
 REFRESH_CONNECTIONS_DIR_LOCK_PATH = (HOME_PROMPT_FLOW_DIR / "refresh_connections_dir.lock").resolve()
 # Note: Use this only for show. Reading input should regard all '*' string as scrubbed, no matter the length.
-SCRUBBED_VALUE = "******"
+SCRUBBED_VALUE = CONNECTION_SCRUBBED_VALUE
 SCRUBBED_VALUE_NO_CHANGE = "<no-change>"
 SCRUBBED_VALUE_USER_INPUT = "<user-input>"
 CHAT_HISTORY = "chat_history"
@@ -143,27 +150,6 @@ SPAN_TABLENAME = "span"
 PFS_MODEL_DATETIME_FORMAT = "iso8601"
 
 AzureMLWorkspaceTriad = namedtuple("AzureMLWorkspace", ["subscription_id", "resource_group_name", "workspace_name"])
-
-
-class CustomStrongTypeConnectionConfigs:
-    PREFIX = "promptflow.connection."
-    TYPE = "custom_type"
-    MODULE = "module"
-    PACKAGE = "package"
-    PACKAGE_VERSION = "package_version"
-    PROMPTFLOW_TYPE_KEY = PREFIX + TYPE
-    PROMPTFLOW_MODULE_KEY = PREFIX + MODULE
-    PROMPTFLOW_PACKAGE_KEY = PREFIX + PACKAGE
-    PROMPTFLOW_PACKAGE_VERSION_KEY = PREFIX + PACKAGE_VERSION
-
-    @staticmethod
-    def is_custom_key(key):
-        return key not in [
-            CustomStrongTypeConnectionConfigs.PROMPTFLOW_TYPE_KEY,
-            CustomStrongTypeConnectionConfigs.PROMPTFLOW_MODULE_KEY,
-            CustomStrongTypeConnectionConfigs.PROMPTFLOW_PACKAGE_KEY,
-            CustomStrongTypeConnectionConfigs.PROMPTFLOW_PACKAGE_VERSION_KEY,
-        ]
 
 
 class RunTypes:
@@ -334,25 +320,6 @@ class ConfigValueType(str, Enum):
     SECRET = "Secret"
 
 
-class ConnectionType(str, Enum):
-    _NOT_SET = "NotSet"
-    AZURE_OPEN_AI = "AzureOpenAI"
-    OPEN_AI = "OpenAI"
-    QDRANT = "Qdrant"
-    COGNITIVE_SEARCH = "CognitiveSearch"
-    SERP = "Serp"
-    AZURE_CONTENT_SAFETY = "AzureContentSafety"
-    FORM_RECOGNIZER = "FormRecognizer"
-    WEAVIATE = "Weaviate"
-    SERVERLESS = "Serverless"
-    CUSTOM = "Custom"
-
-
-class ConnectionAuthMode:
-    KEY = "key"
-    MEID_TOKEN = "meid_token"  # Microsoft Entra ID
-
-
 ALL_CONNECTION_TYPES = set(
     map(lambda x: f"{x.value}Connection", filter(lambda x: x != ConnectionType._NOT_SET, ConnectionType))
 )
@@ -488,3 +455,9 @@ class IdentityKeys(str, Enum):
     USER_IDENTITY = "user_identity"
     RESOURCE_ID = "resource_id"
     CLIENT_ID = "client_id"
+
+
+# Note: Keep these for backward compatibility
+CustomStrongTypeConnectionConfigs = CustomStrongTypeConnectionConfigs
+ConnectionType = ConnectionType
+ConnectionAuthMode = ConnectionAuthMode
