+++ src/promptflow/promptflow/client/__init__.py
@@ -1,0 +1,11 @@
+# ---------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# ---------------------------------------------------------
+__path__ = __import__("pkgutil").extend_path(__path__, __name__)  # type: ignore
+
+# control plane sdk functions
+from promptflow._sdk._load_functions import load_run
+
+from .._sdk._pf_client import PFClient
+
+__all__ = ["PFClient", "load_run"]
