+++ src/promptflow/promptflow/core/__init__.py
@@ -3,21 +3,13 @@
 # ---------------------------------------------------------
 __path__ = __import__("pkgutil").extend_path(__path__, __name__)  # type: ignore
 
-
 from promptflow._core.metric_logger import log_metric
 
 # flake8: noqa
 from promptflow._core.tool import ToolProvider, tool
 
-# control plane sdk functions
-from promptflow._sdk._load_functions import load_flow, load_run
-
-from ._sdk._pf_client import PFClient
-from ._version import VERSION
-
 # backward compatibility
 log_flow_metric = log_metric
 
-__version__ = VERSION
-
-__all__ = ["PFClient", "load_flow", "load_run", "log_metric", "ToolProvider", "tool"]
+# TODO: Add the Flow class
+__all__ = ["log_metric", "ToolProvider", "tool"]
