+++ src/promptflow/tests/sdk_cli_test/unittests/test_utils.py
@@ -36,17 +36,17 @@ from promptflow._sdk._utils import (
     _generate_connections_dir,
     decrypt_secret_value,
     encrypt_secret_value,
+    gen_uuid_by_compute_info,
     generate_flow_tools_json,
+    get_mac_address,
+    get_system_info,
     override_connection_config_with_environment_variable,
     refresh_connections_dir,
     resolve_connections_environment_variable_reference,
-    snake_to_camel,
-    get_mac_address,
-    gen_uuid_by_compute_info,
-    get_system_info,
 )
 from promptflow._utils.load_data import load_data
 from promptflow._utils.retry_utils import http_retry_wrapper, retry
+from promptflow._utils.utils import snake_to_camel
 from promptflow._utils.version_hint_utils import check_latest_version
 
 TEST_ROOT = Path(__file__).parent.parent.parent
